requirements.txt
pandas==2.0.3
pytest==7.4.4
python-dateutil==2.8.2